Helper method to get the repository and make sure it is not in merge state
